Marks the selectors and actions of the commands store as a public API (#51169)
