+++ packages/commands/src/index.js
@@ -2,3 +2,4 @@ export { CommandMenu } from './components/command-menu';
 export { privateApis } from './private-apis';
 export { default as useCommand } from './hooks/use-command';
 export { default as useCommandLoader } from './hooks/use-command-loader';
+export { store } from './store';
