+++ packages/commands/src/store/actions.js
@@ -104,17 +104,3 @@ export function close() {
 		type: 'CLOSE',
 	};
 }
-
-/**
- * Sets the active context.
- *
- * @param {string} context Context.
- *
- * @return {Object} action.
- */
-export function setContext( context ) {
-	return {
-		type: 'SET_CONTEXT',
-		context,
-	};
-}
