+++ packages/edit-site/src/components/site-hub/index.js
@@ -20,7 +20,7 @@ import { store as coreStore } from '@wordpress/core-data';
 import { decodeEntities } from '@wordpress/html-entities';
 import { forwardRef } from '@wordpress/element';
 import { search, external } from '@wordpress/icons';
-import { privateApis as commandsPrivateApis } from '@wordpress/commands';
+import { store as commandsStore } from '@wordpress/commands';
 
 /**
  * Internal dependencies
@@ -28,8 +28,6 @@ import { privateApis as commandsPrivateApis } from '@wordpress/commands';
 import { store as editSiteStore } from '../../store';
 import SiteIcon from '../site-icon';
 import { unlock } from '../../private-apis';
-
-const { store: commandsStore } = unlock( commandsPrivateApis );
 
 const HUB_ANIMATION_DURATION = 0.3;
